Track transport mode in the flow itself
This eliminates the dependency on there having to be a SA,
which is not the case in lazy mode. Set transport from the
policy for lazy mode and set the direction to out.

We're now installing flows, except the peer addresses are
all zeroes. The SADB_ACQUIRE will therefore not match...