Don't add the lazy flow to the active flow tree
Active flows are assumed to have an IKE SA, which lazy flows
do not have. Since even non-active flows are looked up on an
acquire, we actually find the flow. Since it's not an active
flow, we also try to establish a SA. The latter is not quite
working, but were a small step further.