Fix lazy mode
We removed the outgoing flow from the kernel when we closed the SA.
We would never re-add it, because flow_loaded was still set in the
policy.

While we can simply reset the field and have the flow re-inserted,
it would create a blackhole between the removal and a subsequent
re-addition. It's better to avoid removal altogether.