Rename config_findtransform() to config_first_transform()
The function returns the first matching transform. For finding
a particular transform, add config_find_transform().  We'll use
this shortly to look up if the DH group is allowed.

In preparation of that, handle IKEV2_N_INVALID_KE_PAYLOAD in
ikev2_send_informational(). Use config_first_transform() to
suggest a DH group.